coordinator/requirements.txt
# SSUBB Coordinator 依赖
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
httpx>=0.27.0
pyyaml>=6.0
